# bff/docs/swagger.yaml
definitions:
...
  class.AddClassRequest:
    properties:
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      dur_class:
        description: 第几节 '形如 "1-3","1-1"'
        type: string
      name:
        description: 课程名称
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 教师
        type: string
      weeks:
        description: 哪些周
        items:
          type: integer
        type: array
      where:
        description: 地点
        type: string
      year:
        description: 学年
        type: string
    required:
    - day
    - dur_class
    - name
    - semester
    - teacher
    - weeks
    - where
    - year
    type: object
  class.ClassInfo:
    properties:
      class_when:
        description: 上课是第几节（如1-2,3-4）
        type: string
      classname:
        description: 课程名称
        type: string
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      id:
        description: 集合了课程信息的字符串，便于标识（课程ID）
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 任课教师
        type: string
      week_duration:
        description: 上课的周数
        type: string
      weeks:
        description: 哪些周
        items:
          type: integer
        type: array
      where:
        description: 上课地点
        type: string
      year:
        description: 学年
        type: string
    required:
    - class_when
    - classname
    - credit
    - day
    - id
    - semester
    - teacher
    - week_duration
    - weeks
    - where
    - year
    type: object
  class.DeleteClassNoteReq:
    properties:
      classId:
        description: 课程ID
        type: string
      semester:
        description: 学期
        type: string
      year:
        description: 学年
        type: string
    required:
    - classId
    - semester
    - year
    type: object
  class.DeleteClassRequest:
    properties:
      id:
        description: 要被删的课程id
        type: string
      semester:
        description: 学期 "1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        type: string
      year:
        description: 学年  "2024" -> 代表"2024-2025学年"
        type: string
    required:
    - id
    - semester
    - year
    type: object
  class.GetClassListResp:
    properties:
      classes:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
      last_refresh_time:
        description: 上次刷新时间的时间戳,上海时区
        type: integer
    required:
    - classes
    - last_refresh_time
    type: object
  class.GetRecycleBinClassInfosResp:
    properties:
      classInfos:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
    required:
    - classInfos
    type: object
  class.GetSchoolDayResp:
    properties:
      holiday_time:
        type: integer
      school_time:
        type: integer
    required:
    - holiday_time
    - school_time
    type: object
  class.RecoverClassRequest:
    properties:
      classId:
        description: 课程的ID（唯一标识） 更新后这个可能会换，所以响应的时候会把新的ID返回
        type: string
      semester:
        description: 学期 "1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        type: string
      year:
        description: 学年  "2024" 代表"2024-2025学年"
        type: string
    required:
    - classId
    - semester
    - year
    type: object
  class.SearchClassResp:
    properties:
      classInfos:
        items:
          $ref: '#/definitions/class.ClassInfo'
        type: array
    required:
    - classInfos
    type: object
  class.UpdateClassNoteReq:
    properties:
      classId:
        description: 课程ID
        type: string
      note:
        description: 备注
        type: string
      semester:
        description: 学期
        type: string
      year:
        description: 学年
        type: string
    required:
    - classId
    - note
    - semester
    - year
    type: object
  class.UpdateClassRequest:
    properties:
      classId:
        description: 课程的ID（唯一标识） 更新后这个可能会换，所以响应的时候会把新的ID返回
        type: string
      credit:
        description: 学分
        type: number
      day:
        description: 星期几
        type: integer
      dur_class:
        description: 第几节 '形如 "1-3","1-1"'
        type: string
      name:
        description: 课程名称
        type: string
      semester:
        description: 学期
        type: string
      teacher:
        description: 教师
        type: string
      weeks:
        description: 哪些周
        items:
          type: integer
        type: array
      where:
        description: 地点
        type: string
      year:
        description: 学年
        type: string
    required:
    - classId
    - semester
    - year
    type: object
...
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /class/add:
    post:
      consumes:
      - application/json
      description: 添加新的课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 课表信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.AddClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功添加课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 添加课表
      tags:
      - class
  /class/day/get:
    get:
      description: 获取当前周
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取到当前周
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.GetSchoolDayResp'
              type: object
      summary: 获取当前周
      tags:
      - class
  /class/delete:
    post:
      consumes:
      - application/json
      description: 根据课表ID删除课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除课表请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.DeleteClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除课表
      tags:
      - class
  /class/get:
    get:
      description: 根据学期、学年等条件获取课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: refresh
        required: true
        type: boolean
      - in: query
        name: semester
        required: true
        type: string
      - description: 学年,格式为"2024"代表"2024-2025学年"`
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回课表
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.GetClassListResp'
              type: object
      summary: 获取课表
      tags:
      - class
  /class/getRecycle:
    get:
      description: 获取已删除但未彻底清除的课表信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 学期,格式为"1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        in: query
        name: semester
        required: true
        type: string
      - description: 学年,格式为"2024"代表"2024-2025学年"
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取回收站课表信息
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.GetRecycleBinClassInfosResp'
              type: object
      summary: 获取回收站课表信息
      tags:
      - class
  /class/note/delete:
    post:
      consumes:
      - application/json
      description: 根据课程 ID 删除课程备注
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除课程备注请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.DeleteClassNoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除课程备注
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除课程备注
      tags:
      - class
  /class/note/insert:
    post:
      consumes:
      - application/json
      description: 根据课程 ID 更新课程备注
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新课程备注请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.UpdateClassNoteReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功插入课程备注
          schema:
            $ref: '#/definitions/web.Response'
      summary: 插入课程备注
      tags:
      - class
  /class/recover:
    put:
      consumes:
      - application/json
      description: 从回收站恢复课表
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 恢复课表请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.RecoverClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功恢复课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 恢复课表
      tags:
      - class
  /class/search:
    get:
      description: 根据关键词[教师或者课程名]搜索课程,**注意,但当返回的结果数量大于page_size时,代表还有下一页**,最开始请求的是第一页
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        required: true
        type: integer
      - description: 搜索关键词,匹配的是课程名称和教师姓名
        in: query
        name: searchKeyWords
        required: true
        type: string
      - description: 学期,格式为"1"代表第一学期，"2"代表第二学期，"3"代表第三学期
        in: query
        name: semester
        required: true
        type: string
      - description: 学年,格式为"2024"代表"2024-2025学年"
        in: query
        name: year
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功搜索到课程
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/class.SearchClassResp'
              type: object
      summary: 搜索课程
      tags:
      - class
  /class/update:
    put:
      consumes:
      - application/json
      description: 根据课表ID更新课表信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新课表请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/class.UpdateClassRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新课表
          schema:
            $ref: '#/definitions/web.Response'
      summary: 更新课表信息
      tags:
      - class
...
swagger: "2.0"